Starting Application from scratch
