--- src/components/Board/board.css
@@ -1,6 +1,0 @@
-.board-block{
-    width: 25px;
-    height: 25px; 
-    border: 1px solid black;
-    background-color: lightgray;
-}
