--- src/components/Cell/cell.css
@@ -1,20 +1,0 @@
-.cell {
-    width: 25px;
-    height: 25px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    font-size: 16px;
-    cursor: pointer;
-    border: 1px solid #ccc;
-    background-color: #bbb;
-  }
-  
-  .cell.revealed {
-    background-color: #e0e0e0;
-  }
-  
-  .cell.flagged {
-    color: red;
-  }
-
