+++ src/features/cell/Cell.js
@@ -1,0 +1,8 @@
+export const Cell = () => {
+    return (
+        <div>
+            <h1>Hello World</h1>
+        </div>
+    )
+
+}
