+++ src/features/game/Game.js
@@ -1,0 +1,18 @@
+import * as React from 'react';
+import CssBaseline from '@mui/material/CssBaseline';
+import Box from '@mui/material/Box';
+import Container from '@mui/material/Container';
+import { Cell } from '../cell/Cell';
+
+export function Game() {
+  return (
+    <React.Fragment>
+      <CssBaseline />
+      <Container fixed>
+        <Box sx={{ bgcolor: '#cfe8fc', height: '100vh' }}>
+            <Cell/>
+        </Box>
+      </Container>
+    </React.Fragment>
+  );
+}
